Fixed test

+++ InterestCalculatorApi.UnitTests/Controllers/InterestCalculatorControllerTests.cs
@@ -22,7 +22,7 @@ namespace InterestCalculatorApi.UnitTests.Controllers
 
         [Theory]
         [InlineData(100, 5, 105.10)]
-        public void Get_WhenCalled_ReturnsInterestCalculated(double valorinicial, int meses, int jurosResultado)
+        public void Get_WhenCalled_ReturnsInterestCalculated(double valorinicial, int meses, double jurosResultado)
         {
             var getResult = (OkObjectResult) _controller.Get(valorinicial, meses).Result;
             Assert.Equal(jurosResultado, getResult.Value);
